Experimentally, add memory fence
Will need to circle around to this commit and remove the fences that
don't serve any purpose. While investigating reasons for stalls or
hangs, I went a little overboard with the placement of the fences.

But my initial question still stands: as the liburing author added a
fence in their code long after this crate's shared memory accesses were
implemented, should this crate take the same types of precautions or are
those already taken into account and I just missed them?