Modify build.rs to use hdr/io_uring.h
This is still a bit of a hack, as I don't know if this
will be accepted upstream. So the original line
for including linux/io_uring.h is still shown, but commented.

And a builder clang_arg of "-F." is added to get clang
to discover the local header.

The intention is that the src/sys/sys.rs can be built by running

    cargo build --all-features

regardless of the version of Linux the build is happening on.