Update the src/sys/sys.rs
Having run

    cargo build --all-features

with a modified build.rs and a recent hdr/io_uring.h in place,
the sys.rs module can be synced with the most recent liburing
and kernel io_uring version.

As an earlier commit makes the sys module public, this gives access
to io_uring flags to the fork of tokio-uring.